+++ app/src/main/java/com/example/nozes/ProfileActivity.java
@@ -1,0 +1,13 @@
+package com.example.nozes;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class ProfileActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_profile);
+    }
+}
