+++ app/src/main/res/drawable/ic_check_box_desactivated_40dp.xml
@@ -1,0 +1,5 @@
+<vector android:height="40dp" android:tint="#F2F2F2"
+    android:viewportHeight="24.0" android:viewportWidth="24.0"
+    android:width="40dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M19,3L5,3c-1.11,0 -2,0.9 -2,2v14c0,1.1 0.89,2 2,2h14c1.11,0 2,-0.9 2,-2L21,5c0,-1.1 -0.89,-2 -2,-2zM10,17l-5,-5 1.41,-1.41L10,14.17l7.59,-7.59L19,8l-9,9z"/>
+</vector>
