+++ app/src/main/res/drawable/ic_person_desactivated_40dp.xml
@@ -1,0 +1,5 @@
+<vector android:height="40dp" android:tint="#F2F2F2"
+    android:viewportHeight="24.0" android:viewportWidth="24.0"
+    android:width="40dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M12,12c2.21,0 4,-1.79 4,-4s-1.79,-4 -4,-4 -4,1.79 -4,4 1.79,4 4,4zM12,14c-2.67,0 -8,1.34 -8,4v2h16v-2c0,-2.66 -5.33,-4 -8,-4z"/>
+</vector>
