+++ app/src/main/res/drawable/shape_button_nav.xml
@@ -1,0 +1,11 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <solid
+        android:color="#F2F2F2"/>
+    <padding
+        android:top="10dp"
+        android:right="10dp"
+        android:bottom="10dp"
+        android:left="10dp"/>
+    <corners android:radius="10dp"/>
+</shape>
